Support D3D 'result_bytebuffer' case
D3D output parameters with a SAL labl including the text
'result_bytebuffer' and a size appear similar to arrays of bytes but are
really a pointer to memory allocated internally by the implementation
with the specified size. For this case, the following adjustments are
made for capture and replay, where capture will continue to record the
content of the retrieved memory and replay will allocate and retrieve a
single pointer to the memory.
- Update encoding for this case, where the type is void**, to
  dereference the pointer when invoking EncodeVoidArray.
- Update replay to decode the array of bytes, but allocate a single
  pointer parameter for API call replay, using PointerDecoder<uint8_,
  void*> to represent the parameter.