Prevent array size null pointer dereference
Update code generation to add a null check when dereferencing
a pointer to an array size when encoding array data during capture and
allocating array memory during replay. Added for cases where both the
variable length array parameter and its associated size parameter are
optional and are allowed to be null.