Avoid opaque pointer copies
When making deep copies of structures, avoid attempting to copy the
values referenced by opaque handles and opaque pointers. These are
generally pointers to a single value with the void* type, which cannot
be safely copied because the size of the referenced value is unknown.
For most cases, the referenced value is unneeded because capture treats
the address stored in these handles/pointers as external object IDs
that are encoded as 64-bit integer values.

One exception, which is not handled by this change, is the pData value
from the VkPushDescriptorSetWithTemplateInfoKHR struct. This struct
needs a custom copy routine to process pData based on the layout
defined at template creation.